requirements.txt
flask
python-dotenv
sqlalchemy
psycopg2-binary
msal
openpyxl
requests
pytest
flask-limiter
